backend/requirements.txt
# EduStory Requirements

# Core Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# AI/LLM Integration
langchain==0.1.0
openai==1.7.2
anthropic==0.8.1  # For Claude if needed

# Image Generation
stability-sdk==0.8.4
pillow==10.1.0

# HTTP Client
httpx==0.25.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment Management
python-dotenv==1.0.0

# Async Support
aiofiles==23.2.1

# Database (for production)
sqlalchemy==2.0.23
alembic==1.13.1

# Monitoring & Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx-mock==0.15.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.7.1
